Add some eye candy to show the user when a page cannot load.  Fix a bug in the controller that was still paying attention to the screen number

+++ DynamicDisplays/src/gov/fnal/ppd/dd/display/client/ConnectionToBrowserInstance.java
@@ -488,6 +488,6 @@ public abstract class ConnectionToBrowserInstance {
 
 	protected void errorSeen(long value) {
 		for (BrowserErrorListener L : listeners)
-			L.errorInPageLoad(value);
+			L.errorInPageLoad((int) value);
 	}
 }
